# docs/swagger.yaml
basePath: /
definitions:
  service.FileInfo:
    properties:
      blob_id:
        type: integer
      category:
        type: string
      compression_alg:
        type: string
      content:
        description: Base64 encoded
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      hash:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      old_cumulus_id:
        type: integer
      size_compressed:
        type: integer
      size_raw:
        type: integer
      subtype:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Cumulus3 object storage.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cumulus3 API
  version: "1.0"
paths:
  /base/files/delete/{id}:
    delete:
      description: Deletes a file by its ID
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: File deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a file
      tags:
      - files
  /base/files/id/{id}:
    get:
      description: Downloads a file by its old Cumulus ID
      parameters:
      - description: Old Cumulus ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Download a file by old ID
      tags:
      - files
  /base/files/info/{id}:
    get:
      description: Get detailed information about a file by its old Cumulus ID
      parameters:
      - description: Old Cumulus ID
        in: path
        name: id
        required: true
        type: integer
      - description: Include base64 content
        in: query
        name: extended
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.FileInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get file info by old ID
      tags:
      - files
  /v2/files/{id}:
    get:
      description: Downloads a file by its ID
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Download a file
      tags:
      - files
  /v2/files/info/{id}:
    get:
      description: Get detailed information about a file
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      - description: Include base64 content
        in: query
        name: extended
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.FileInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get file info
      tags:
      - files
  /v2/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file to the storage
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Tags like array of string or coma separated strings
        in: formData
        name: tags
        type: string
      - description: Legacy ID
        in: formData
        name: old_cumulus_id
        type: integer
      - description: Validity period (e.g. '1 day', '2 months')
        in: formData
        name: validity
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "413":
          description: File too large
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload a file
      tags:
      - files
swagger: "2.0"